Pull out glossary logic into a custom hook
