+++ src/app/components/elements/TrustedMarkdown.tsx
@@ -1,19 +1,10 @@
-import React, {useState} from "react";
-import ReactDOMServer from "react-dom/server";
-import {Provider, useSelector, useStore} from "react-redux";
-import * as RS from "reactstrap";
-import {Router} from "react-router-dom";
-import {AppState} from "../../state/reducers";
-import {EXAM_BOARD_NULL_OPTIONS, MARKDOWN_RENDERER} from "../../services/constants";
+import React from "react";
+import {MARKDOWN_RENDERER} from "../../services/constants";
 import {TrustedHtml} from "./TrustedHtml";
-import {IsaacGlossaryTerm} from "../content/IsaacGlossaryTerm";
-import {GlossaryTermDTO} from "../../../IsaacApiTypes";
 import {escapeHtml, replaceEntities} from "remarkable/lib/common/utils";
 import {Token} from "remarkable";
-import uuid from "uuid";
-import {history} from "../../services/history";
-import {useUserContext} from "../../services/userContext";
 import {SITE, SITE_SUBJECT} from "../../services/siteConstants";
+import {useGlossaryTermsInMarkdown} from "../../services/glossary";
 
 MARKDOWN_RENDERER.renderer.rules.link_open = function(tokens: Token[], idx/* options, env */) {
     let href = escapeHtml(tokens[idx].href || "");
@@ -26,85 +17,8 @@ MARKDOWN_RENDERER.renderer.rules.link_open = function(tokens: Token[], idx/* opt
     }
 };
 
-function getTermFromCandidateTerms(candidateTerms: GlossaryTermDTO[]) {
-    if (candidateTerms.length === 0) {
-        return null;
-    } else if (candidateTerms.length === 1) {
-        return candidateTerms[0];
-    } else {
-        console.warn('More than one candidate term was found: ', candidateTerms);
-        return candidateTerms[0];
-    }
-}
-
 export const TrustedMarkdown = ({markdown}: {markdown: string}) => {
-    const store = useStore();
-    const {examBoard} = useUserContext();
-    const examBoardTag = !EXAM_BOARD_NULL_OPTIONS.has(examBoard) ? examBoard.toLowerCase() : "";
-
-    const glossaryTerms = useSelector((state: AppState) => state && state.glossaryTerms);
-    const [componentUuid, setComponentUuid] = useState(uuid.v4().slice(0, 8));
-
-    // This tooltips array is necessary later on: it will contain
-    // UncontrolledTooltip elements that cannot be pre-rendered as static HTML.
-    const tooltips: JSX.Element[] = [];
-
-    // Matches strings such as [glossary:glossary-demo|boolean-algebra] which MUST be at the beginning of the line.
-    // This is used to render the full version of a glossary term using the IsaacGlossaryTerm component.
-    const glossaryBlockRegexp = /^\[glossary:(?<id>[a-z-|]+?)\]/gm;
-
-    // Matches strings such as [glossary-inline:glossary-demo|boolean-algebra] and
-    // [glossary-inline:glossary-demo|boolean-algebra "boolean algebra"] which CAN be inlined.
-    // This is used to produce a hoverable element showing the glossary term, and its definition in a tooltip.
-    const glossaryInlineRegexp = /\[glossary-inline:(?<id>[a-z-|]+?)\s*(?:"(?<text>[A-Za-z0-9 ]+)")?\]/g;
-
-    const glossaryIdsInMarkdown = Array.from(new Set([
-        ...Array.from(markdown.matchAll(glossaryBlockRegexp)).filter(m => m.groups && m.groups.id),
-        ...Array.from(markdown.matchAll(glossaryInlineRegexp)).filter(m => m.groups && m.groups.id),
-    ]));
-
-    if (glossaryTerms && glossaryTerms.length > 0 && glossaryIdsInMarkdown.length > 0) {
-        const filteredTerms = glossaryTerms.filter(term => term.examBoard === "" || term.examBoard === examBoard);
-
-        // Markdown can't cope with React components, so we pre-render our component to static HTML, which Markdown will then ignore.
-        // This requires a bunch of stuff to be passed down along with the component.
-        markdown = markdown.replace(glossaryBlockRegexp, (_match, id) => {
-            const term = getTermFromCandidateTerms(filteredTerms.filter(term => (term.id as string) === id || (term.id as string) === `${id}|${examBoardTag}`));
-            if (term === null) {
-                console.error('No valid term for "' + id + '" found among the filtered terms: ', filteredTerms);
-                return "";
-            }
-
-            return ReactDOMServer.renderToStaticMarkup(
-                <Provider store={store}>
-                    <Router history={history}>
-                        <IsaacGlossaryTerm doc={term}/>
-                    </Router>
-                </Provider>
-            );
-        });
-
-        // This is easier: we replace an inline glossary term with a <span> which is later targeted by ReactStrap's UncontrolledTooltip.
-        // The tooltip components can be rendered as regular react objects, so we just add them to an array,
-        // and return them inside the JSX.Element that is returned as TrustedMarkdown.
-        markdown = markdown.replace(glossaryInlineRegexp, (_match, id, text, offset) => {
-            const term = getTermFromCandidateTerms(filteredTerms.filter(term => (term.id as string) === id || (term.id as string) === `${id}|${examBoardTag}`));
-            if (term === null) {
-                console.error('No valid term for "' + id + '" found among the filtered terms: ', filteredTerms);
-                return "";
-            }
-
-            const cssFriendlyTermId = (term.id as string).replace(/\|/g, '-');
-            const tooltipTargetId = `glossary-${componentUuid}-${cssFriendlyTermId}-${offset}`;
-            // This is properly horrible but it works...
-            tooltips.push(
-                <RS.UncontrolledTooltip placement="bottom" target={tooltipTargetId}>
-                    <TrustedMarkdown markdown={term.explanation && term.explanation.value || ''} />
-                </RS.UncontrolledTooltip>
-            );
-            return `<span class="inline-glossary-term" id="${tooltipTargetId}">${text || term.value}</span>`;
-        });
-    }
+    const [markdownWithGlossary, glossaryTooltips] = useGlossaryTermsInMarkdown(markdown);
 
     // RegEx replacements to match Latex inspired Isaac Physics functionality
     const regexRules = {
@@ -116,7 +30,7 @@ export const TrustedMarkdown = ({markdown}: {markdown: string}) => {
             "[**Concepts**](/concepts)": /\*\*Concepts\*\*/g,
         });
     }
-    let regexProcessedMarkdown = markdown;
+    let regexProcessedMarkdown = markdownWithGlossary;
     Object.entries(regexRules).forEach(([replacement, rule]) =>
         regexProcessedMarkdown = regexProcessedMarkdown.replace(rule, replacement)
     );
@@ -124,6 +38,6 @@ export const TrustedMarkdown = ({markdown}: {markdown: string}) => {
     const html = MARKDOWN_RENDERER.render(regexProcessedMarkdown);
     return <div>
         <TrustedHtml html={html} />
-        {tooltips}
+        {glossaryTooltips}
     </div>;
 };
